Rework core extensions to reconcile class definitions with directory structure
