+++ lib/core_ext/numeric.rb
@@ -1,13 +1,21 @@
-class Numeric
-  def round_to_nearest(rounding_quotient = 0)
-    if rounding_quotient.zero?
-      round
-    else
-      (self / rounding_quotient.to_f).round * rounding_quotient
+# frozen_string_literal: true
+
+module CoreExt
+  module NumericExtensions
+    def round_to_nearest(rounding_quotient = 0)
+      if rounding_quotient.zero?
+        round
+      else
+        (self / rounding_quotient.to_f).round * rounding_quotient
+      end
+    end
+
+    def numericize # Parallel to String#numericize
+      self
     end
   end
+end
 
-  def numericize # Parallel to String#numericize
-    self
-  end
+class Numeric
+  include CoreExt::NumericExtensions
 end
